Fixed some small errors

+++ SolarPowerCalc/src/com/qut/spc/model/converter/SolarPanelConverter.java
@@ -14,7 +14,7 @@ public class SolarPanelConverter extends SolarComponentConverter{
 	}
 	
 	@XmlElement
-	public float getOutputEnergy(){
+	public double getOutputEnergy(){
 		return ((SolarPanel)panel).getOutputEnergy();
 	}
